lift/index.html

<html>
  <head>
    <link href="../tester/testharness.css" rel="stylesheet">
    <script src="../tester/testharness.js"></script>
  </head>
  <body>
    <style>
      body{
        margin: 0;
      }

      .wrapper{
        max-width: 1024px;
        margin: 0 auto;
        position: relative;
        height: 100%;
      }

      .liftCabine{
        border: 4px solid black;
        width: 150px;
        height: 200px;
        position: absolute;
        /*left: 50%;*/
        transform: translateX(-50%);
        transition: bottom 2s ease-in-out;
        -webkit-transition: bottom 2s ease-in-out;
        bottom: 0;
      }

      .liftCabine1{
        left: 40%;
        background-color: red;
      }

      .liftCabine2{
        left: 60%;
        background-color: green;
      }

      .liftCabine.btn-0{
        bottom: -190px;
        transition: bottom 2s ease-in-out;
        -webkit-transition: bottom 2s ease-in-out;
      }

      .liftCabine.btn-1{
        bottom: 0;
        transition: bottom 2s ease-in-out;
        -webkit-transition: bottom 2s ease-in-out;
      }


      .liftCabine.btn-2{
        bottom: 50%;
        transition: bottom 2s ease-in-out;
        -webkit-transition: bottom 2s ease-in-out;
      }


      .liftCabine.btn-3{
        bottom: calc(100% - 210px);
      }

      .buttons, .buttons2{
        width: 50px;
        position: absolute;
        right: 0;
        top: 50%;
        transform: translateY(-50%);
      }

      .buttons1{
        right: 60px;
      }

      .buttons button, .buttons2 button{
        width: 50px;
        height: 50px;
        margin-bottom: 10px;
      }
    </style>
    <div class='wrapper'>
      <div class='liftCabine liftCabine1'>1</div>
      <div class='liftCabine liftCabine2'>2</div>
      <div class='buttons buttons1'>
        1 lift
        <button class='btn-3'>3</button>
        <button class='btn-2'>2</button>
        <button class='btn-1'>1</button>
        <button class='btn-0'>-1</button>
      </div>
      <div class='buttons buttons2'>
        2 lift
        <button class='btn-3'>3</button>
        <button class='btn-2'>2</button>
        <button class='btn-1'>1</button>
        <button class='btn-0'>-1</button>
      </div>
    </div>

<!--lift programme -->
  <script>
    var buttons1 = document.querySelectorAll('.buttons.buttons1 button');
    var buttons2 = document.querySelectorAll('.buttons.buttons2 button');
    var liftCabine1 = document.querySelector('.liftCabine1');

    for (var i = 0; i < buttons1.length; i++){
      buttons1[i].addEventListener('click', function(){
        liftCabine1.classList.add('column1');
      });
    }

  </script>

<!--testing-->
  <script>

;(function() {
    //test if massive of two buttons created
    var buttons1 = document.querySelectorAll('.buttons.buttons1 button');
    var buttons2 = document.querySelectorAll('.buttons.buttons2 button');
    var cabine1 = document.querySelector('.liftCabine1');

    var twoBtnArrayTest = async_test('testing if massive of two buttons created');

    window.addEventListener('load', twoBtnArrayTest.step_func(function(){
      var amount = 4;
      assert_equals(buttons1.length, amount);
      assert_equals(buttons2.length, amount);
      this.done();
    }));

    // test column1
    function callTest(i) {
      var testBtn = async_test('if i click on btn ' + i + ' of first column then lift1 should have class column1');

      testBtn.step(function(){
        var btn = buttons1[i];
        btn.addEventListener('click', testBtn.step_func(function(e){
          // cabine1.classList.add(e.target.className);
          assert_true(cabine1.classList.contains('column1'));
          console.log(e.target);
          testBtn.done();

          // tear-down
          cabine1.classList.remove('column1');
        }));

        btn.dispatchEvent(new Event('click'));
      });
    }

    for (var i = 0; i < buttons1.length; i++) {
      callTest(i);
    }

})();
  </script>
  </body>
</html>
